+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * added regions
 * updated User-Agent blacklist
+* updated referrer blacklist
 * updated referrer mapping
 * updated dependencies